Rename subnavigation component
